requirements.txt
asgiref==3.5.2
cffi==1.15.1
cryptography==39.0.2
Django==4.0.5
django-cors-headers==3.13.0
django-environ==0.10.0
djangorestframework==3.13.1
djangorestframework-jwt==1.11.0
Pillow==9.1.1
pycparser==2.21
PyJWT==2.6.0
python-dateutil==2.8.2
pytz==2022.2.1
six==1.16.0
sqlparse==0.4.2
tzdata==2022.1
urllib3==1.26.6
